requirements.txt
Django==1.7
pyjade==3.0.0
django-crispy-forms==1.4.0
